snn streaming moved to nnet_snn_stream.h
